+++ ImmortalCoordinator/Program.cs
@@ -96,7 +96,7 @@ namespace CRA.Worker
                 dataProvider, descriptor, connectionsPoolPerWorker);
 
             worker.DisableDynamicLoading();
-            worker.SideloadVertex(new AmbrosiaRuntime(), "ambrosia");
+            worker.SideloadVertex(new AmbrosiaNonShardedRuntime(), "ambrosia");
 
             worker.Start();
         }
